# internal/docs/swagger.yaml
definitions:
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.FieldValueInput:
    properties:
      field_id:
        type: string
      value: {}
    type: object
...
  api.GetProjectFieldsResponse:
    properties:
      fields:
        items:
          $ref: '#/definitions/api.projectFieldValueResponse'
        type: array
    type: object
...
  api.PatchProjectFieldsRequest:
    properties:
      values:
        items:
          $ref: '#/definitions/api.FieldValueInput'
        type: array
    type: object
...
  api.projectFieldValueResponse:
    properties:
      definition_deleted:
        type: boolean
      field_id:
        type: string
      field_type:
        type: string
      key:
        type: string
      name:
        type: string
      value: {}
    type: object
info:
  contact: {}
paths:
...
  /api/v1/projects/{id}/fields:
    get:
      description: Returns all custom field values set on the project. Only fields
        with <code>scope=project</code> are included. The response structure mirrors
        the asset fields endpoint.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetProjectFieldsResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Project not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get project field values
      tags:
      - Custom Fields
    patch:
      consumes:
      - application/json
      description: Sets or clears one or more custom field values on the project.
        Only fields with <code>scope=project</code> are accepted — submitting an asset-scoped
        field returns 422.<br><br> Pass <code>null</code> as the value to clear a
        field. Returns the full updated field values for the project.
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Field values to set
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.PatchProjectFieldsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GetProjectFieldsResponse'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Project or field not found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Value validation failed or wrong scope
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update project field values
      tags:
      - Custom Fields
...
swagger: "2.0"